# target/generated-sources/openapi/src/main/java/com/bank/yanki/api/YankiApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bank.yanki.api;

import com.bank.yanki.model.CreateWalletRequest;
import com.bank.yanki.model.LinkDebitCardRequest;
import com.bank.yanki.model.TopupRequest;
import com.bank.yanki.model.WalletResponse;
import com.bank.yanki.model.WithdrawalRequest;
import com.bank.yanki.model.YankiPaymentRequest;
import com.bank.yanki.model.YankiPaymentResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "yanki", description = "the yanki API")
public interface YankiApi {

    /**
     * POST /yanki/payments : Enviar pago P2P entre monederos por número de celular
     *
     * @param yankiPaymentRequest  (required)
     * @return Pago realizado (status code 201)
     *         or Monedero no encontrado (status code 404)
     *         or Regla de negocio no cumplida (status code 422)
     */
    @Operation(
        operationId = "createPayment",
        summary = "Enviar pago P2P entre monederos por número de celular",
        responses = {
            @ApiResponse(responseCode = "201", description = "Pago realizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = YankiPaymentResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Monedero no encontrado"),
            @ApiResponse(responseCode = "422", description = "Regla de negocio no cumplida")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/yanki/payments",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<YankiPaymentResponse>> createPayment(
        @Parameter(name = "YankiPaymentRequest", description = "", required = true) @Valid @RequestBody YankiPaymentRequest yankiPaymentRequest
    );


    /**
     * POST /yanki/wallets : Crear un monedero Yanki
     *
     * @param createWalletRequest  (required)
     * @return Monedero creado (status code 201)
     *         or Solicitud inválida (status code 400)
     *         or Regla de negocio no cumplida (status code 422)
     */
    @Operation(
        operationId = "createWallet",
        summary = "Crear un monedero Yanki",
        responses = {
            @ApiResponse(responseCode = "201", description = "Monedero creado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = WalletResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida"),
            @ApiResponse(responseCode = "422", description = "Regla de negocio no cumplida")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/yanki/wallets",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<WalletResponse>> createWallet(
        @Parameter(name = "CreateWalletRequest", description = "", required = true) @Valid @RequestBody CreateWalletRequest createWalletRequest
    );


    /**
     * GET /yanki/wallets/{id} : Obtener monedero por ID
     *
     * @param id  (required)
     * @return Monedero encontrado (status code 200)
     *         or Monedero no encontrado (status code 404)
     */
    @Operation(
        operationId = "getWalletById",
        summary = "Obtener monedero por ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Monedero encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = WalletResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Monedero no encontrado")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/yanki/wallets/{id}",
        produces = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<WalletResponse>> getWalletById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * GET /yanki/wallets/by-phone/{phoneNumber} : Obtener monedero por número de celular
     *
     * @param phoneNumber  (required)
     * @return Monedero encontrado (status code 200)
     *         or Monedero no encontrado (status code 404)
     */
    @Operation(
        operationId = "getWalletByPhone",
        summary = "Obtener monedero por número de celular",
        responses = {
            @ApiResponse(responseCode = "200", description = "Monedero encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = WalletResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Monedero no encontrado")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/yanki/wallets/by-phone/{phoneNumber}",
        produces = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<WalletResponse>> getWalletByPhone(
        @Parameter(name = "phoneNumber", description = "", required = true, in = ParameterIn.PATH) @PathVariable("phoneNumber") String phoneNumber
    );


    /**
     * POST /yanki/wallets/{walletId}/link-debit-card : Vincular monedero a una tarjeta de débito del banco
     *
     * @param walletId  (required)
     * @param linkDebitCardRequest  (required)
     * @return Tarjeta vinculada (status code 200)
     *         or Monedero no encontrado (status code 404)
     *         or Regla de negocio no cumplida (status code 422)
     */
    @Operation(
        operationId = "linkDebitCard",
        summary = "Vincular monedero a una tarjeta de débito del banco",
        responses = {
            @ApiResponse(responseCode = "200", description = "Tarjeta vinculada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = WalletResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Monedero no encontrado"),
            @ApiResponse(responseCode = "422", description = "Regla de negocio no cumplida")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/yanki/wallets/{walletId}/link-debit-card",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<WalletResponse>> linkDebitCard(
        @Parameter(name = "walletId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("walletId") String walletId,
        @Parameter(name = "LinkDebitCardRequest", description = "", required = true) @Valid @RequestBody LinkDebitCardRequest linkDebitCardRequest
    );


    /**
     * GET /yanki/wallets : Listar monederos Yanki
     *
     * @return Lista de monederos (status code 200)
     */
    @Operation(
        operationId = "listWallets",
        summary = "Listar monederos Yanki",
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de monederos", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = WalletResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/yanki/wallets",
        produces = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<List<WalletResponse>>> listWallets(
        
    );


    /**
     * POST /yanki/topups : Fondear monedero desde la cuenta principal asociada a la tarjeta
     *
     * @param topupRequest  (required)
     * @return Fondeo realizado (status code 201)
     *         or Monedero no encontrado (status code 404)
     *         or Regla de negocio no cumplida (status code 422)
     */
    @Operation(
        operationId = "topup",
        summary = "Fondear monedero desde la cuenta principal asociada a la tarjeta",
        responses = {
            @ApiResponse(responseCode = "201", description = "Fondeo realizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = WalletResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Monedero no encontrado"),
            @ApiResponse(responseCode = "422", description = "Regla de negocio no cumplida")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/yanki/topups",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<WalletResponse>> topup(
        @Parameter(name = "TopupRequest", description = "", required = true) @Valid @RequestBody TopupRequest topupRequest
    );


    /**
     * POST /yanki/wallets/{walletId}/unlink-debit-card : Desvincular monedero de la tarjeta de débito
     *
     * @param walletId  (required)
     * @return Tarjeta desvinculada (status code 200)
     *         or Monedero no encontrado (status code 404)
     */
    @Operation(
        operationId = "unlinkDebitCard",
        summary = "Desvincular monedero de la tarjeta de débito",
        responses = {
            @ApiResponse(responseCode = "200", description = "Tarjeta desvinculada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = WalletResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Monedero no encontrado")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/yanki/wallets/{walletId}/unlink-debit-card",
        produces = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<WalletResponse>> unlinkDebitCard(
        @Parameter(name = "walletId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("walletId") String walletId
    );


    /**
     * POST /yanki/withdrawals : Retirar saldo del monedero hacia la cuenta principal asociada a la tarjeta
     *
     * @param withdrawalRequest  (required)
     * @return Retiro realizado (status code 201)
     *         or Monedero no encontrado (status code 404)
     *         or Regla de negocio no cumplida (status code 422)
     */
    @Operation(
        operationId = "withdraw",
        summary = "Retirar saldo del monedero hacia la cuenta principal asociada a la tarjeta",
        responses = {
            @ApiResponse(responseCode = "201", description = "Retiro realizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = WalletResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Monedero no encontrado"),
            @ApiResponse(responseCode = "422", description = "Regla de negocio no cumplida")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/yanki/withdrawals",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<WalletResponse>> withdraw(
        @Parameter(name = "WithdrawalRequest", description = "", required = true) @Valid @RequestBody WithdrawalRequest withdrawalRequest
    );

}